chore: refactor manager into file
